% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_test.R
\name{convolution_test}
\alias{convolution_test}
\title{Identify driver genes}
\usage{
convolution_test(bmr, n)
}
\arguments{
\item{bmr}{A Backround Mutation Rate (BMR) matrix. Rows and columns correspond to genes and patients.}

\item{n}{A mutation count matrix. Rows and columns correspond to genes and patients.}
}
\value{
A vector of p values
}
\description{
\code{convolution_test} assigns a p value for each gene.
}
\examples{
# Load the example dataset:
data(mutation)

convolution_test(bmr, n)

}
\author{
Weiqing Huang
}
