% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBMR.R
\name{ProBMR}
\alias{ProBMR}
\title{Poisson Regression of Background Mutation Rate (BMR)}
\usage{
ProBMR(m, M, x)
}
\arguments{
\item{m}{A mutation count matrix. Rows and columns correspond to genes and patients.}

\item{M}{A vector of gene coverage.}

\item{x}{A genomic covariates matrix. Rows and columns correspond to genes and features.}
}
\value{
A list with components:
\describe{
\item{bkgd1}{A list contains bmr (fitted values), fixef (fixed effects), ranef (random effects), VarCorr (Variance and Correlation Components) in the first regression.}
\item{bkgd2}{A list contains bmr (fitted values), ranef (random effects), VarCorr (Variance and Correlation Components) in the second regression.}
}
}
\description{
\code{ProBMR} estimates BMRs using two mixed-effects Poisson regression models applied to link gene associated features and accommodate the overdispersion in mutation counts.
}
\examples{
# Load the example dataset:
data(mutation)

fit <- ProBMR(m, M, x)

}
\author{
Weiqing Huang
}
